Move VotingCard to component folder.
