+++ together-wherever/src/components/cards/VotingCard.tsx
